+++ keyboards/40percentclub/nano/keymaps/drashna/rules.mk
@@ -3,3 +3,5 @@ RGBLIGHT_ENABLE          	= no
 CONSOLE_ENABLE 				= no
 
 BOOTLOADER 					= qmk-dfu
+
+SRC += analog.c
