+++ keyboards/kyria/keymaps/drashna/config.h
@@ -37,3 +37,5 @@
 // If you are using an Elite C rev3 on the slave side, uncomment the lines below:
 #define SPLIT_USB_DETECT
 #define SPLIT_USB_TIMEOUT 1000
+
+#define KEYLOGGER_LENGTH 10
