+++ keyboards/kyria/keymaps/drashna/rules.mk
@@ -20,3 +20,5 @@ HD44780_ENABLE = no        # Enable support for HD44780 based LCDs
 BOOTLOADER = atmel-dfu
 
 SPLIT_TRANSPORT = mirror
+
+TAP_DANCE_ENABLE = yes
